Define setup.sh and reorganize location of container images.
The purpose of this commit is to define a setup.sh file for developers
to execute prior to starting development. This setup, for now, just
builds the container images used during the execution of tests.

The functions in `gittest` package no longer attempt to build the image
on each execution run and simply relies on the developer to perform the
building of the images prior to development.

Additionally, to support the tests to run during CI, the ci.yml file was
updated to execute the setup.sh file created.