Provide ability to get license from repository root.
The purpose of this commit is to provide the ability to retrieve the
license for a repository given if a LICENSE file exists at the root of
the repository.

To determine the license the package go-license-detector is used. This
package returns a map of license strings as the key and confidence
value as the values. To determine the license, the license with the
highest confidence value is returned.

If no LICENSE file exists at the root then we return NO LICENSE.