Remove outside package type object.Commit in RepoDetails' AuthorsCommits
The purpose of this commit is to remove the use of the object.Commit
type provided by go-git. Instead, a new struct, Commit, is defined that
takes a subset of the details provided by the object.Commit type. This
is to reduce the dependency of any consumers of RepoReader so they do
not know the underlying existence of go-git when receiving a
RepoDetails response or the now map[Author][]Commit type.