Provide ability to have the tui.EntryModel have RepoReader dependency
The purpose of this commit is to ultimately allow for the tui.EntryModel
to receive the ReaderRepo as a dependency.

During the course of this work the NewApp constructor was refactored to
receive a repo directory path as an argument rather than a
tui.EntryModel. This is so the cmd package does not need to know about
the underlying tui and should only be concerned with managing cli
related tasks and using the App to call validation and appropriate
calls.

As part of these arguments from the cmd, we want to verify that at most
1 arg is provided. If no args are provided we assume the working
directory should be evaluated. If an arg is provided we should verify
that it is indeed a directory. If it is not a directory then the
RepoReader shouldn't need to consider it. If it is a directory then the
RepoReader may consider if a git repository is in it.