Refactor RepoReader.GetAuthorsByCommits
The purpose of this commit is to simply refactor
RepoReader.GetAuthorsByCommits to return a map with the key value as
the author's email instead of the Author type. This will resolve any
discrepancies where an author goes by a pseudonym or multiple
pseudonyms as we will identify them by their email.

For display purposes in the Overview of the TUI, we will simply obtain
the name from the last commit in the slice of commits for the author.