Create Repository With Authors Having Multiple Aliases
The purpose of this commit is to provide a repository of multiple
authors that have made commits using different names but under the same
email. That is, the repository has multiple authors that go by
different pseudonyms.

For example, Author2 Alias1 has the email gitcha2@gitcha.com. They also
go by Author2 Alias2 and so should have their commits collected under
the email.